move `FuncResponse`, `CommandResponse` etc from `coreutils` to `federation` package
