+++ pkg/coreutils/federation/types.go
@@ -22,3 +22,35 @@ type implIFederation struct {
 }
 
 type OffsetsChan chan istructs.Offset
+
+// implements json.Unmarshaler
+type CommandResponse struct {
+	NewIDs            map[string]istructs.RecordID
+	CurrentWLogOffset istructs.Offset
+	CmdResult         map[string]interface{}
+}
+
+type QueryResponse struct {
+	QPv2Response // TODO: eliminate this after https://github.com/voedger/voedger/issues/1313
+	Sections     []struct {
+		Elements [][][][]interface{} `json:"elements"`
+	} `json:"sections"`
+}
+
+type QPv2Response []map[string]interface{}
+
+func (r QPv2Response) Result() map[string]interface{} {
+	return r.ResultRow(0)
+}
+
+func (r QPv2Response) ResultRow(rowNum int) map[string]interface{} {
+	return r[rowNum]
+}
+
+// implements json.Unmarshaler
+type FuncResponse struct {
+	coreutils.HTTPResponse
+	CommandResponse
+	QueryResponse
+	SysError error
+}
